requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0